Make package version != Chromedriver version.
Forcing package version (a semantic version in the form major.minor.patch) to mirror the Chromedriver public version (in form major.minor) was limiting. It forced some unnecessary complex logic, and didn’t allow for a new package release unless a new version of Chromedriver came out. Keeping them separate makes more sense.

The ”version” property of the main export is now a String with the version of the Chromedriver binary.