Use fewer dependencies.
The old version was only dependent on two other npm dependencies BUT they
used 1000* dependencies. This version will make installation faster.

We also skip throwing an error when installing the driver. This makes it possible for Browsertime users to only use Firefox or install Chromedriver themself for non supported systems.